openapi: Fix number -> integer issues

+++ client/model_cluster_profile_google_google_node_pools_pool1.go
@@ -12,9 +12,9 @@ package client
 
 type ClusterProfileGoogleGoogleNodePoolsPool1 struct {
 	Autoscaling bool `json:"autoscaling,omitempty"`
-	Count float32 `json:"count,omitempty"`
-	MinCount float32 `json:"minCount,omitempty"`
-	MaxCount float32 `json:"maxCount,omitempty"`
+	Count int32 `json:"count,omitempty"`
+	MinCount int32 `json:"minCount,omitempty"`
+	MaxCount int32 `json:"maxCount,omitempty"`
 	InstanceType string `json:"instanceType,omitempty"`
 	ServiceAccount string `json:"serviceAccount,omitempty"`
 }
